hr/team-profiles/requirements-dev.txt
# Test dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
responses>=0.24.0

# Runtime dependencies (for testing imports)
pyyaml>=6.0.0
requests>=2.31.0
